añadido trabajo final

+++ controllers/uploads.js
@@ -1,18 +1,19 @@
 const path = require('path');
-const fs   = require('fs');
+const fs= require('fs');
 
 
 const { response } = require('express');
 const { uploadFile } = require('../helpers/uploadFile');
 
-const { User } = require('../models/usuario');
+const categoria = require('../models/categoria')
+const producto=require('../models/producto')
 const { request } = require('http');
-const cerveza = require('../models/cerveza');
+const usuario = require('../models/usuario');
 
 
 
 
-const upload = async(req, res = response) => {
+const upload = async(req=request, res = response) => {
     
     if (!req.files || Object.keys(req.files).length === 0) {
         res.status(400).send('No files were uploaded.');
@@ -36,28 +37,33 @@ const updateImage=async (req=request,res=response)=>{
     const {id,collection}=req.params;
     let objeto;
     switch (collection) {
-        case "cerveza":
-             objeto=await cerveza.findById(id);
+        case "categoria":
+             objeto=await categoria.findById(id);
             
             break;
-            case "users":
-                objeto=await cerveza.findById(id);
+            case "producto":
+                objeto=await producto.findById(id);
+                break;
+            case "usuarios":
+                objeto=await usuario.findById(id);
                 break;
         default:
-            res.json({
+           return res.json({
                 message:"No es una colección valida"
             })
             break;
            
     }
     if (objeto){
-        const ruta=path.join(__dirname,'../uploads',collection,objeto.img);
-        if (!fs.existsSync(ruta)){
-            return res.status(400).json({
-                message:"No existe la imagen"
-            })
-        }
+        if (objeto.img){
+            const ruta=path.join(__dirname,'../uploads',collection,objeto.img);
+            if (!fs.existsSync(ruta)){
+                return res.status(400).json({
+                    message:"No existe la imagen"
+                })
+            }
             fs.unlinkSync(ruta);
+        }
             const nombre = await uploadFile( req.files,['png','jpg','jpeg','gif'] , collection );
             objeto.img=nombre
             await objeto.save();
@@ -65,6 +71,10 @@ const updateImage=async (req=request,res=response)=>{
                 filename:nombre
             })
 
+    }else {
+        res.json({
+            message:"No tiene imagen"
+        })
     }
